Add happy path for creating new SurveyRespondent with manual POST request

+++ app/models/SurveyRespondent.scala
@@ -1,5 +1,5 @@
 package models
 
-case class SurveyRespondent(id: Option[Long] = None, username: String, survey_id: Long = 1) {
+case class SurveyRespondent(id: Option[Long] = None, username: String, survey_id: Int = 1) {
 
 }
